plugin: Add ability to start/stop perf ctrs using signals
This patch introduces two new concepts:

 - ability to start/stop perf counters by sending the process
   signals from any external entity.

 - ability to control the behavior of the plugin using two new
   environ variables: DMTCP_START_CTRS_ON_RESTART_STRATEGY, and
   DMTCP_KILL_ON_RESUME_STRATEGY.

The changes have been done keeping the future requirements
in mind. In particular, this will allow an external entity to
dynamically change the behavior of the plugin, and sample perf
statistics at a time of its choosing, without having to modify the
plugin source.