parallel_dmtcp: Add template code for soft deadline
If a running job finishes in the middle of a time epoch, we start a new
random job, provided that the running job finished early enough, i.e.,
before a hard deadline.

This patch introduces the concept of a "soft" deadline. If a running job
finishes after the soft deadline but before the hard deadline, we pick
restart a random job from our pool, and extend the duration of the time
epoch for all the running jobs, including the new job.

The code for the soft deadline is commented out for now. It will be
enabled after testing in a follow-up patch.